+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Output a data processor as a dashboard.
 * Output a data processor as a tab on the contact summary screen.
 * Output a data processor as a contribution search.
+* Output a data processor as a membership search.
 * Added field outputs for simple calculations (substract and total).
 * Added escaped output to search screens.
 * Replaced the value separator in the raw field with a comma.